# PERF: assert_frame_equal / assert_series_equal