- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.

Follow-up to #55949 addressing some of @jbrockmendel's comments and also expanding to non-multiindex cases:

```
import pandas as pd

N = 1_000_000

idx = pd._testing.makeStringIndex(N)

ser1 = pd.Series(1, index=idx.copy(deep=True))
ser2 = pd.Series(1, index=idx.copy(deep=True))

%timeit pd.testing.assert_series_equal(ser1, ser2)

# 17.8 s ± 129 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# 48.1 ms ± 3.28 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```